+++ public/change_menu_button_event.js
@@ -1,5 +1,5 @@
 export class ChangeMenuButtonEvent extends CustomEvent {
-    constructor(style) {
-        super('MenuButton:change', {detail: {style: style}});
+    constructor(detail) {
+        super('MenuButton:change', {detail: detail});
     }
 }