+++ public/menu_extensions.js
@@ -51,7 +51,7 @@ export class MenuExtensions extends Menu {
                 let optionValue = String(
                     this.core.getOption(
                         this.core.idToName(
-                            item.id.slice((this.menuId + '-oed-').length)
+                            item.id.slice((this.id + '-oed-').length)
                         )
                     )
                 );