+++ public/menu_file.js
@@ -139,7 +139,7 @@ export class MenuFile extends Menu {
                 const optionValue = String(
                     this.core.getOption(
                         this.core.idToName(
-                            item.id.slice((this.menuId + '-oed-').length)
+                            item.id.slice((this.id + '-oed-').length)
                         )
                     )
                 );