+++ public/menu_help.js
@@ -80,7 +80,7 @@ export class MenuHelp extends Menu {
                 const optionValue = String(
                     this.core.getOption(
                         this.core.idToName(
-                            item.id.slice((this.menuId + '-oed-').length)
+                            item.id.slice((this.id + '-oed-').length)
                         )
                     )
                 );